Do not update report date time unless it parking violation
