+++ app/src/main/java/lt/vilnius/tvarkau/mvp/presenters/NewReportPresenter.kt
@@ -1,6 +1,7 @@
 package lt.vilnius.tvarkau.mvp.presenters
 
 import lt.vilnius.tvarkau.utils.FieldAwareValidator
+import java.io.File
 
 /**
  * @author Martynas Jurkus
@@ -14,4 +15,6 @@ interface NewReportPresenter {
     fun initWithReportType(reportType: String)
 
     fun submitProblem(validator: FieldAwareValidator<NewReportData>)
+
+    fun onImagesPicked(imageFiles: List<File>)
 }
