+++ app/src/main/java/lt/vilnius/tvarkau/mvp/views/NewReportView.kt
@@ -2,6 +2,7 @@ package lt.vilnius.tvarkau.mvp.views
 
 import lt.vilnius.tvarkau.entity.Profile
 import lt.vilnius.tvarkau.utils.FieldAwareValidator
+import java.io.File
 
 /**
  * @author Martynas Jurkus
@@ -19,4 +20,8 @@ interface NewReportView {
     fun showError(error: Throwable)
 
     fun showSuccess()
+
+    fun fillReportDateTime(dateTime: String)
+
+    fun displayImages(imageFiles: List<File>)
 }
